+++ lua/obsidian/lsp/init.lua
@@ -7,6 +7,9 @@ local log = require "obsidian.log"
 ---@return integer?
 lsp.start = function(buf)
   local capabilities = vim.lsp.protocol.make_client_capabilities()
+  capabilities.workspace = capabilities.workspace or {}
+  capabilities.workspace.fileOperations =
+    vim.tbl_extend("force", capabilities.workspace.fileOperations or {}, { didRename = true })
 
   local lsp_config = {
     name = "obsidian-ls",