+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prevent crash when dumping nested empty arrays.
 - Mapping `q` to be quit in normal mode in snacks picker.
 - Wrong migration process for `opts.use_advanced_uri`.
+- Tag finding will not generate false positives of other list fields.
 
 ## [v3.16.0](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.16.0) - 2026-03-16