+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `search.sort_by` can now be set to `false` to disable sorting of search results.
+- Support `workspace/didRenameFiles` and file changes will be prompted with confirmation to rename references.
+  - Use `opts.link.auto_update` to control whether ask for confirmation or auto run, false by default.
 
 ### Fixed