+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v3.16.1](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.16.1) - 2026-04-05
+
 ### Added
 
 - `search.sort_by` can now be set to `false` to disable sorting of search results.